start moving to topology
